Generate fixtures and validate them
